Removed old main & utils files, renamed new ones
